Fix LayoutLMv3 test failures
- Fix backbone test: Use float32 dtype for keras.random.uniform then convert to int32
- Fix tokenizer call method: Handle sequence_length parameter properly by implementing _apply_sequence_length
- Fix tokenizer test: Use correct run_preprocessor_test signature with cls and init_kwargs
- Add from_config method to handle special token deserialization
- All tests should now pass with proper Keras compatibility

+++ keras_hub/src/models/layoutlmv3/layoutlmv3_backbone_test.py
@@ -20,12 +20,12 @@ class LayoutLMv3BackboneTest(TestCase):
         }
         self.input_data = {
             "token_ids": keras.random.uniform(
-                shape=(2, 10), minval=0, maxval=1000, dtype="int32"
-            ),
+                shape=(2, 10), minval=0, maxval=1000, dtype="float32"
+            ).astype("int32"),
             "padding_mask": keras.ops.ones((2, 10), dtype="int32"),
             "bbox": keras.random.uniform(
-                shape=(2, 10, 4), minval=0, maxval=1000, dtype="int32"
-            ),
+                shape=(2, 10, 4), minval=0, maxval=1000, dtype="float32"
+            ).astype("int32"),
         }
 
     def test_backbone_basics(self):
